+++ examples/test-site/src/components/FilterByGroup/index.tsx
@@ -67,11 +67,6 @@ const asFilterByGroup = flow(
 const FilterByGroup = asFilterByGroup(FilterByGroupClean);
 
 export default FilterByGroup;
-
 export {
   ContextLogger,
 } from './ContextLogger';
-
-export {
-  AccordionFilter,
-} from './AccordionFilter';
