+++ examples/test-site/src/data/pages/filter-page/filter$default$category_name.json
@@ -1,0 +1,3 @@
+{
+  "text": "New Category"
+}
