+++ examples/test-site/src/data/pages/filter-page/filter$default.json
@@ -1,0 +1,3 @@
+{
+  "on": true
+}
